+++ python/GafferSceneTest/__init__.py
@@ -202,6 +202,7 @@ from .PrimitiveQueryTest import PrimitiveQueryTest
 from .SceneStatsTest import SceneStatsTest
 from .PointInstancerAlgoTest import PointInstancerAlgoTest
 from .CopyObjectTest import CopyObjectTest
+from .QuantizePrimitiveVariablesTest import QuantizePrimitiveVariablesTest
 
 from .IECoreScenePreviewTest import *
 from .IECoreGLPreviewTest import *